[INTERNAL][I] Split LocalEditorStatusChangeHandler into separate classes
Splits LocalEditorStatusChangeHandler into separate classes to better
differentiate between the different responsibilities.

Introduces the abstract parent class
AbstractLocalEditorStatusChangeHandler to avoid duplicating code needed
to set up each new handler.

Groups the handlers in EditorManager by the resource they are handling
to improve the overview of which handlers are used.

Adjusts the context test mock setup to handle multiple calls
instantiating a MessageBusConnection. This is needed as there now are
four classes that create a MessageBusConnection instead of only one.
